Extract engine subtitle burn-in operation
subtitles is a focused burn-in operation with FFmpeg filter escaping and output probing. This moves it into engine_subtitles.py while re-exporting subtitles from mcp_video.engine for compatibility.

Constraint: Preserve subtitle/style escaping, subtitles filter requirement, and EditResult behavior.

Rejected: Move generate_subtitles with subtitles | generated SRT creation has a different contract and should be reviewed separately.

Confidence: high

Scope-risk: narrow

Directive: Keep security-sensitive FFmpeg filter operations small and explicitly verified.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_subtitles.py

Tested: /opt/homebrew/bin/python3 subtitles re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'subtitles' tests/test_server.py -k 'subtitles' tests/test_cli.py -k 'subtitles' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
